Split out language picker from header
