+++ src/components/Header/Header.js
@@ -4,11 +4,9 @@ import { Link } from 'react-router';
 import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
 
-export default class Header extends React.Component {
-  changeLanguage(event, index, value) {
-    this.props.flux.actions.DASHBOARD.changeLanguage(value);
-  }
+import LanguagePicker from './LanguagePicker';
 
+export default class Header extends React.Component {
   render() {
     return (
       <nav className="navbar navbar-trans" role="navigation">
@@ -79,20 +77,12 @@ export default class Header extends React.Component {
   }
 
   renderLanguagePicker() {
-    const languages = this.props.supportedLanguages.map(lang =>
-      <MenuItem key={lang} value={lang} primaryText={lang} />
-    );
-
     return (
-      <SelectField
-        underlineStyle={{ borderColor: '#337ab7', borderBottom: 'solid 3px' }}
-        labelStyle={{ fontWeight: 600, color: '#2ebd59' }}
-        value={this.props.language}
-        autoWidth={true}
-        style={{maxWidth:'60px'}}
-        onChange={(event, index, value) => this.changeLanguage(event, index, value)}>
-          {languages}
-      </SelectField>
+      <LanguagePicker
+        supportedLanguages={this.props.supportedLanguages}
+        language={this.props.language}
+        flux={this.props.flux}
+      />
     );
   }
 };
